+++ apps/playground/tsconfig.app.json
@@ -2,6 +2,7 @@
   "extends": "@vue/tsconfig/tsconfig.dom.json",
   "include": ["env.d.ts", "src/**/*", "src/**/*.vue"],
   "compilerOptions": {
+    "noEmit": false,
     "composite": true,
     "baseUrl": ".",
     "types": ["vite/client"]
